updated redirects
